Re-duplicate hue_rotate_color
